Normalize paths in error messages before comparing them to snapshots
The path pointing to any given file or directory that's a child of
the directory containing the Tyson repository is going to be different
on every device, due to the facts that 1) each developer could have
the Tyson repository at a different location on their device, and 2)
different devices use different path separators (POSIX "/" vs. Windows
"\").

As a real life example, as of this commit, snapshot testing currently
fails on CI because "/Users/kyle/tyson/test/fixtures/" doesn't match
"/home/travis/build/kylejlin/tyson/test/fixtures/".